requirement.txt
fastapi
uvicorn
SQLAlchemy
async-exit-stack
async-generator
aiofiles
jinja2
pytest
requests
PyYAML
fastapi-utils